Consolidate getIncludeData instances
